+++ src/simplification/detail/Quadrics.h
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "algorithms/simplification/QEMSimplifier.h"
+#include "algorithms/simplification/SimplificationTypes.h"
+#include "core/Mesh.h"
 #include "detail/FeatureGuidance.h"
 #include "detail/SimplificationTypes.h"